[refactor] the editable form is now a modal and in the read only view there is a resuming table
