+++ css/styles.scss
@@ -193,3 +193,10 @@ textarea {
     font-size: 24px;
 }
 
+.form-group {
+    margin-bottom: 0rem !important;
+}
+
+button {
+    margin: 5px !important;
+}
